refactor: facade 인터페이스 제거
